+++ src/pipeline_telemetry/storage/__init__.py
@@ -1,1 +1,2 @@
+from .generic import AbstractTelemetryStorage  # noqa
 from .mongo import TelemetryMongoStorage  # noqa