+++ tests/test_decorator.py
@@ -1,8 +1,11 @@
 """Module to test decorator logic.
 """
+import pytest
 from test_data import DEFAULT_TELEMETRY_PARAMS
 
-from pipeline_telemetry.decorator import add_mongo_telemetry, add_telemetry
+from pipeline_telemetry.decorator import add_mongo_single_usage_telemetry, \
+    add_mongo_telemetry, add_single_usage_telemetry, add_telemetry
+from pipeline_telemetry.settings import exceptions
 from pipeline_telemetry.storage.memory import TelemetryInMemoryStorage
 from pipeline_telemetry.storage.mongo import TelemetryMongoStorage
 
@@ -111,3 +114,76 @@ def test_calling_mongo_decorated_method_from_within_decorated_method():
     assert not hasattr(class_instance, "_telemetry")
     assert class_instance.decorated_method() == DEFAULT_TELEMETRY_PARAMS.get("category")
     assert hasattr(class_instance, "_telemetry")
+
+
+def test_single_usage_decorator_raises_params_not_def_exc():
+    """
+    Test that add_single_usage_telemetry decorator raises an exception when
+    no telemetry params are defined in class of class instance.
+    """
+
+    class DecoratorTest:
+        @add_single_usage_telemetry()
+        def decorated_method(self):
+            return "method result"
+    with pytest.raises(exceptions.ClassTelemetryParamsNotDefined) as excep:
+        DecoratorTest().decorated_method()
+
+    assert 'Telemetry params not defined' in str(excep)
+
+
+def test_single_usage_class_telemetry_settings():
+    """
+    Test that add_single_usage_telemetry decorator can use class defined
+    telemetry settings.
+    """
+
+    class DecoratorTest:
+        TELEMETRY_PARAMS = DEFAULT_TELEMETRY_PARAMS
+
+        @add_single_usage_telemetry()
+        def decorated_method(self):
+            return "method result"
+
+    class_instance = DecoratorTest()
+    assert not hasattr(class_instance, "_telemetry")
+    assert class_instance.decorated_method() == "method result"
+    assert hasattr(class_instance, "_telemetry")
+
+
+def test_single_usage_telemetry_settings_with_sub_process():
+    """
+    Test that add_single_usage_telemetry decorator can use class defined
+    telemetry settings.
+    """
+    sub_process = DEFAULT_TELEMETRY_PARAMS['process_type'].sub_processes[0]
+
+    class DecoratorTest:
+        TELEMETRY_PARAMS = DEFAULT_TELEMETRY_PARAMS
+
+        @add_mongo_single_usage_telemetry(sub_process=sub_process)
+        def decorated_method(self):
+            return self._telemetry.telemetry
+
+    class_instance = DecoratorTest()
+
+    telemetry = class_instance.decorated_method()
+    assert telemetry.get(sub_process).get('base_counter') == 1
+
+
+def test_single_usage_telemetry_settings_with_mongho_storage(mocker):
+    """
+    Test that mongo decorator sets the mongo storage class
+    """
+    mongo_module_path = "pipeline_telemetry.storage.mongo."
+    mocker.patch(mongo_module_path + "TelemetryMongoModel.save",
+                 return_value=None)
+
+    class DecoratorTest:
+        TELEMETRY_PARAMS = DEFAULT_TELEMETRY_PARAMS
+
+        @add_mongo_single_usage_telemetry()
+        def decorated_method(self):
+            return self._telemetry.storage_class
+
+    assert isinstance(DecoratorTest().decorated_method(), TelemetryMongoStorage)